Refactor: Unify Control Panel and add new tutorial
This commit refactors the application to unify several separate views into a single "Panel de Control" dashboard.

The new "Panel de Control" now includes:
- Key Performance Indicators (KPIs) for ECRs and ECOs.
- Charts showing the status distribution of ECRs and ECOs.
- The ECR log for tracking departmental approvals.
- The meeting attendance matrix and absenteeism charts.

The previous, separate views (`indicadores_ecm_view`, `ecr_seguimiento`, `seguimiento_ecr_eco`) have been removed, and the navigation has been updated accordingly.

A new, specific tutorial (`tutorialControlPanel.js`) has been created for the unified dashboard, and a button has been added to launch it. The old, more general tutorial has been removed.
