+++ public/index.html
@@ -160,10 +160,6 @@
                             <div class="border-t my-1"></div>
                             <a href="#" data-view="control_ecrs" class="nav-link flex items-center gap-3 px-4 py-3 text-sm hover:bg-slate-100"><i data-lucide="layout-dashboard" class="w-5 h-5 text-slate-500"></i>Panel de Control</a>
                             <a href="#" data-view="seguimiento_ecr_eco" class="nav-link flex items-center gap-3 px-4 py-3 text-sm hover:bg-slate-100"><i data-lucide="file-check" class="w-5 h-5 text-slate-500"></i>Seguimiento ECR/ECO</a>
-                            <div class="border-t my-1"></div>
-                            <a href="#" id="start-tutorial-btn" class="nav-link flex items-center gap-3 px-4 py-3 text-sm hover:bg-slate-100 text-blue-600 font-semibold">
-                                <i data-lucide="life-buoy" class="w-5 h-5"></i>Iniciar Tutorial
-                            </a>
                         </div>
                     </div>
                     <!-- Dropdown de Gestión -->
@@ -246,7 +242,6 @@
 -->
 <!-- Firebase Hosting Scripts -->
 <script type="module" src="main.js"></script>
-<script type="module" src="tutorial.js"></script>
 
 </body>
 </html>
